backend/requirements.txt
asgiref==3.4.1
Django==3.2.15
django-filter==21.1
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
drf-extra-fields==3.1.1
fpdf==1.7.2
gunicorn==20.1.0
isort==5.10.1
Pillow==9.0.1
psycopg2-binary==2.9.2
pytz==2021.3
reportlab==3.6.3
sqlparse==0.4.2
python-dotenv==0.20.0
djoser==2.1.0

